[FEAT] CLI mostly completed

+++ src/stt/fractor.rs
@@ -272,8 +272,8 @@ impl Fractor {
             }
         }
 
-        //
-        if erase {
+        // Erase the temp directory
+        if erase && output_dir.exists() {
             fs::remove_dir_all(output_dir).with_context(|| "Unable to remove temp audio files")?;
         }
